requirements.txt
eel
Pillow
plyer
io
pystray